Relocate webhook parsers
