+++ src/internal/webhook-handler/__tests__/webhook-handler.subs.test.ts
@@ -3,7 +3,7 @@ import {
     parseChannelSubscriptionGiftsEvent,
     parseChannelSubscriptionNewEvent,
     parseChannelSubscriptionRenewalEvent
-} from '../webhook-handler';
+} from '../webhook-parsers';
 
 describe('parseChannelSubscriptionNewEvent', () => {
     beforeAll(() => {
